Homework 7: Networking
* add: retrofit

* add: pagination

* add: room database

* add: fix issues

+++ app/src/main/java/com/example/homeworkAA/extensions/Exts.kt
@@ -1,15 +1,13 @@
 package com.example.homeworkAA.extensions
 
 import android.os.Bundle
-import com.example.homeworkAA.data.models.Movie
-import java.lang.IllegalArgumentException
 
 private const val KEY_MOVIE_ID = "KEY_MOVIE_ID"
 
-var Bundle?.movieIdBundle: Int
+var Bundle?.movieIdBundle: Long
     get() {
-        return this?.getInt(KEY_MOVIE_ID) ?: throw IllegalArgumentException("No movie in bundle")
+        return this?.getLong(KEY_MOVIE_ID) ?: throw IllegalArgumentException("No movie in bundle")
     }
     set(value) {
-        this?.putInt(KEY_MOVIE_ID, value)
+        this?.putLong(KEY_MOVIE_ID, value)
     }
